requirements.txt
pandas
pyyaml
google-adk
pydantic
openai
python-dotenv